requirements.txt
blue==0.9.1
Django==5.0.2
flake8==4.0.1
isort==5.13.2
mypy==1.7.1
pylint==3.0.3
taskipy==1.12.2
pytest==8.0.1
pytest-django==4.8.0
psycopg2-binary==2.9.9
pydantic==2.6.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
ipdb==0.13.13